+++ www/gatsby-config.js
@@ -95,7 +95,7 @@ const plugins = [
     resolve: 'gatsby-plugin-page-creator',
     options: {
       path: `${__dirname}/src/pages`,
-      ignore: ['tools/usage-insights.tsx'],
+      ignore: ['insights.tsx'],
     },
   },
 ];
