+++ www/gatsby-node.js
@@ -116,7 +116,7 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
   INSIGHTS_PAGES.forEach(({ path: pagePath, tab }) => {
     createPage({
       path: pagePath,
-      component: require.resolve('./src/pages/tools/usage-insights.tsx'),
+      component: require.resolve('./src/pages/insights.tsx'),
       context: { tab },
     });
   });
