+++ www/src/pages/tools/component-generator.mdx
@@ -76,7 +76,7 @@ If you wish to use SASS variables (which is the preferred way of styling the com
 easily overridden and customized by the consumers of Paragon), add them in `/src/MyComponent/_variables.scss` (this file should contain all variables specific to your component).
 This way the variables will also get automatically picked up by documentation site and displayed on your component's page.
 
-**Please note that you need to follow [Paragon's CSS styling conventions](docs/decisions/0012-css-styling-conventions).**
+**Please note that you need to follow [Paragon's CSS styling conventions](https://github.com/openedx/paragon/blob/master/docs/decisions/0012-css-styling-conventions).**
 
 ## 5. Document your component
